Make each bootstrapper file its own module. This allows us to import them all directly into the bootstrapper, instead of having a separate bootstrapApp file.
