+++ bootstrapper/inputs/inputBootstrapper.ts
@@ -6,6 +6,8 @@ import { services } from 'typescript-angular-utilities';
 import __timezone = services.timezone;
 import __validation = services.validation;
 
+export const moduleName: string = 'InputTestModule';
+
 interface ITestItem {
 	name: string;
 }
@@ -76,5 +78,5 @@ class InputTestController {
 	}
 }
 
-angular.module('app')
+angular.module(moduleName, [])
 	.controller('InputTestController', InputTestController);
