+++ bootstrapper/text/text.ts
@@ -1,5 +1,7 @@
 import * as angular from 'angular';
 
+export const moduleName: string = 'TextModule';
+
 const textComponent: angular.IComponentOptions = {
 	transclude: true,
 	template: '<span>{{$ctrl.text}}</span>',
@@ -19,5 +21,5 @@ class TextController {
 	}
 }
 
-angular.module('app')
+angular.module(moduleName, [])
 	.component('tsText', textComponent);
